+++ backend/server/src/main/java/com/todoc/server/domain/realtime/web/controller/WebSocketAuthHandshakeInterceptor.java
@@ -20,6 +20,9 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class WebSocketAuthHandshakeInterceptor implements HandshakeInterceptor {
 
+    /**
+     * WebSocket 세션 연결 전에 요청 정보 파싱 및 인증 절차 수행
+     */
     @Override
     public boolean beforeHandshake(ServerHttpRequest req,
                                    ServerHttpResponse res,
@@ -47,7 +50,6 @@ public class WebSocketAuthHandshakeInterceptor implements HandshakeInterceptor {
             return false;
         }
 
-        // TODO: 인증 토큰으로 userId 등 추가
         SessionAuth auth = (SessionAuth) session.getAttribute("AUTH_USER");
         long authId = auth.id();
